Updated ascripts

+++ ASCRIPT_INSTALL.py
@@ -2,17 +2,19 @@ from subprocess import call
 import subprocess
 import json
 import configparser
+import os
 
-firebaseProjectID = 'Your Project ID HERE'
+firebaseProjectID = 'amtannotate3'
 
-config = configparser.RawConfigParser()
-config.add_section('Username Here')
-config.add_section('SetUp')
-config.set('SetUp', 'firebaseSubdomain', firebaseProjectID)
-config.set('Username Here', 'awskey', 'Your AWS Key Here')
-config.set('Username Here', 'awssakey', 'Your AWS Secret Access Key')
-with open('config.ini', 'w') as configfile:
-    config.write(configfile)
+if not (os.path.exists('config.ini')):
+	config = configparser.RawConfigParser()
+	config.add_section('Username Here')
+	config.add_section('SetUp')
+	config.set('SetUp', 'firebaseSubdomain', firebaseProjectID)
+	config.set('Username Here', 'awskey', 'Your AWS Key Here')
+	config.set('Username Here', 'awssakey', 'Your AWS Secret Access Key')
+	with open('config.ini', 'w') as configfile:
+	    config.write(configfile)
 
 todump = {}
 todump["projects"] = {}
@@ -21,4 +23,7 @@ json.dump(todump, open('toWeb/.firebaserc', 'w'))
 def subprocess_cmd(command):
     process = subprocess.Popen(command,stdout=subprocess.PIPE, shell=True)
     proc_stdout = process.communicate()[0].strip()
-    print (p
+    print (proc_stdout)
+print('Deploying Site... This may take a few minutes')
+subprocess_cmd ('cd toWeb && firebase deploy ')
+print('Site Deployed! Installation Finished. Please insert your AWS Access and Secret Access Keys inside the config file')
